scripts/requirements-smartsheet.txt
smartsheet-python-sdk>=2.177.0
python-dotenv>=1.0.0
